[Clean Code] Make a fading link to avoid history push in many places
PATH=/definition/metier
